widget-examples/database-connectors/snowflake_connector_python/requirements.txt
snowflake-connector-python
python-dotenv
fastapi
pandas
uvicorn